refactor: reorganize OCI-related code and update environment variable handling
