+++ .dagger/common/constants.go
@@ -1,0 +1,11 @@
+package common
+
+// These constants are duplicated in internal/pkg/testhelpers/constants.go.
+// Make sure to update both files when changing these values.
+const (
+	IntegrationTestingVarName = "INTEGRATION_TEST"
+	RegistryHostVarName       = "REGISTRY_HOST"
+	RegistryAuthHostVarName   = "REGISTRY_AUTH_HOST"
+	RegistryUsernameVarName   = "REGISTRY_USERNAME"
+	RegistryPasswordVarName   = "REGISTRY_PASSWORD"
+)
