+++ internal/pkg/testhelpers/constants.go
@@ -1,5 +1,7 @@
-package shared
+package testhelpers
 
+// These constants are duplicated in .dagger/common/constants.go.
+// Make sure to update both files when changing these values.
 const (
 	IntegrationTestingVarName = "INTEGRATION_TEST"
 	RegistryHostVarName       = "REGISTRY_HOST"
